move authentication logic to auth.js
